--- a/src/main/java/kr/co/pinup/posts/model/dto/PostResponse.java
+++ b/src/main/java/kr/co/pinup/posts/model/dto/PostResponse.java
@@ -8,19 +8,37 @@ import java.time.LocalDateTime;
 
 @Builder
 public record PostResponse(Long id, Long storeId, MemberResponse member, String title, String content,
-                           String thumbnail, LocalDateTime createdAt, LocalDateTime updatedAt,  int commentCount
-                          , boolean likedByCurrentUser ) {
+                           String thumbnail, LocalDateTime createdAt, LocalDateTime updatedAt
+                          ,  int commentCount, int  likeCount, boolean likedByCurrentUser ) {
 
     public static PostResponse from(Post post) {
         return new PostResponse(post.getId(), post.getStore().getId(), new MemberResponse(post.getMember()),
                 post.getTitle(), post.getContent(), post.getThumbnail(),
-                post.getCreatedAt(), post.getUpdatedAt(), 0, false);
+                post.getCreatedAt(), post.getUpdatedAt(), 0, 0, false);
     }
 
-    public static PostResponse fromPostWithComments(Post post, int commentCount, boolean likedByCurrentUser) {
-        return new PostResponse(post.getId(), post.getStore().getId(), new MemberResponse(post.getMember()),
-                post.getTitle(), post.getContent(), post.getThumbnail(),
-                post.getCreatedAt(), post.getUpdatedAt(), commentCount, likedByCurrentUser);
+    public PostResponse(Long id,
+                        String memberNickname,
+                        String title,
+                        String thumbnail,
+                        LocalDateTime createdAt,
+                        Long commentCount,
+                        Integer likeCount,
+                        Boolean likedByCurrentUser) {
+        this(
+                id,
+                null,
+                MemberResponse.ofNickname(memberNickname),
+                title,
+                 null,
+                thumbnail,
+                createdAt,
+                null,
+                commentCount != null ? commentCount.intValue() : 0,
+                likeCount != null ? likeCount.intValue() : 0,
+                Boolean.TRUE.equals(likedByCurrentUser)
+        );
     }
+
 }