--- a/src/main/java/kr/co/pinup/posts/repository/PostRepository.java
+++ b/src/main/java/kr/co/pinup/posts/repository/PostRepository.java
@@ -2,6 +2,7 @@ package kr.co.pinup.posts.repository;
 
 import jakarta.persistence.LockModeType;
 import kr.co.pinup.posts.Post;
+import kr.co.pinup.posts.model.dto.PostResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
@@ -22,4 +23,31 @@ public interface PostRepository extends JpaRepository<Post, Long> {
     @Query("SELECT p FROM Post p WHERE p.id = :id")
     Optional<Post> findByIdWithOptimisticLock(@Param("id") Long id);
 
+    @Query("""
+            SELECT new kr.co.pinup.posts.model.dto.PostResponse(
+                p.id,
+                p.member.nickname,
+                p.title,
+                p.thumbnail,
+                p.createdAt,
+                COUNT(DISTINCT c.id),
+                p.likeCount,
+              CASE
+                WHEN :memberId IS NOT NULL AND
+                     EXISTS (
+                       SELECT 1 FROM PostLike pl
+                       WHERE pl.post.id = p.id AND pl.member.id = :memberId
+                     )
+                THEN TRUE ELSE FALSE
+              END
+            )
+            FROM Post p
+            LEFT JOIN Comment c ON c.post.id = p.id
+            LEFT JOIN PostLike pl ON pl.post.id = p.id AND pl.member.id = :memberId
+            WHERE p.store.id = :storeId AND p.isDeleted = :isDeleted
+            GROUP BY p.id, p.store.id, p.member.nickname, p.title, 
+                     p.thumbnail, p.createdAt, p.updatedAt, p.likeCount
+            ORDER BY p.createdAt DESC
+            """)
+    List<PostResponse> findPostListItems(@Param("storeId") Long storeId, @Param("isDeleted") boolean isDeleted, @Param("memberId") Long memberId);
 }