--- a/src/main/java/kr/co/pinup/posts/repository/PostRepository.java
+++ b/src/main/java/kr/co/pinup/posts/repository/PostRepository.java
@@ -30,23 +30,23 @@ public interface PostRepository extends JpaRepository<Post, Long> {
                 p.title,
                 p.thumbnail,
                 p.createdAt,
-                COUNT(DISTINCT c.id),
+                COUNT(c.id),
                 p.likeCount,
-              CASE
-                WHEN :memberId IS NOT NULL AND
-                     EXISTS (
-                       SELECT 1 FROM PostLike pl
-                       WHERE pl.post.id = p.id AND pl.member.id = :memberId
-                     )
-                THEN TRUE ELSE FALSE
-              END
+                CASE
+                  WHEN :memberId IS NOT NULL AND
+                       EXISTS (
+                         SELECT 1 FROM PostLike pl
+                         WHERE pl.post.id = p.id AND pl.member.id = :memberId
+                       )
+                  THEN TRUE ELSE FALSE
+                END
             )
             FROM Post p
             LEFT JOIN Comment c ON c.post.id = p.id
-            LEFT JOIN PostLike pl ON pl.post.id = p.id AND pl.member.id = :memberId
-            WHERE p.store.id = :storeId AND p.isDeleted = :isDeleted
-            GROUP BY p.id, p.store.id, p.member.nickname, p.title, 
-                     p.thumbnail, p.createdAt, p.updatedAt, p.likeCount
+            WHERE p.store.id = :storeId
+              AND p.isDeleted = :isDeleted
+            GROUP BY
+              p.id, p.member.nickname, p.title, p.thumbnail, p.createdAt, p.likeCount
             ORDER BY p.createdAt DESC
             """)
     List<PostResponse> findPostListItems(@Param("storeId") Long storeId, @Param("isDeleted") boolean isDeleted, @Param("memberId") Long memberId);