--- a/src/main/java/kr/co/pinup/posts/service/PostService.java
+++ b/src/main/java/kr/co/pinup/posts/service/PostService.java
@@ -33,6 +33,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -97,30 +99,13 @@ public class PostService {
     public List<PostResponse> findByStoreIdWithCommentsAndLikes(Long storeId, boolean isDeleted, MemberInfo memberInfo) {
         log.debug("댓글 포함 게시글 요청: storeId={}, isDeleted={}", storeId, isDeleted);
 
-        List<Post> posts = postRepository.findByStoreIdAndIsDeleted(storeId, isDeleted);
+        Long memberId = (memberInfo != null) ? memberRepository.findByNickname(memberInfo.nickname())
+                .orElseThrow(() -> new MemberNotFoundException(memberInfo.nickname() + "님을 찾을 수 없습니다."))
+                .getId() : null;
 
-        if (memberInfo == null) {
-            return posts.stream()
-                    .map(post -> PostResponse.fromPostWithComments(
-                            post,
-                            commentRepository.countByPostId(post.getId()),
-                            false
-                    ))
-                    .collect(Collectors.toList());
-        }
+        return postRepository.findPostListItems(storeId, isDeleted, memberId);
 
-        Member member = memberRepository.findByNickname(memberInfo.nickname())
-                .orElseThrow(() -> new MemberNotFoundException(memberInfo.nickname() + "님을 찾을 수 없습니다."));
-
-        return posts.stream()
-                .map(post -> PostResponse.fromPostWithComments(
-                        post,
-                        commentRepository.countByPostId(post.getId()),
-                        postLikeRepository.existsByPostIdAndMemberId(post.getId(), member.getId())
-                ))
-                .collect(Collectors.toList());
     }
-
 
     public PostResponse getPostById(Long id, boolean isDeleted) {
         log.debug("게시글 단건 요청: postId={}, isDeleted={}", id, isDeleted);
@@ -231,7 +216,6 @@ public class PostService {
             throw e;
         }
     }
-
 
     public void disablePost(Long postId) {
         Post post = findByIdOrThrow(postId);