--- a/src/main/resources/static/css/post_list.css
+++ b/src/main/resources/static/css/post_list.css
@@ -1,182 +1,130 @@
+/* ========== Top meta ========== */
 .wrap_profile {
     display: flex;
     align-items: center;
     gap: 5px;
     font-size: 14px;
     color: #555;
-    margin-bottom: 10px;
+    margin-bottom: 5px;
 }
+.post_content_wrap { margin: 0 auto; padding: 20px 0; }
 
+/* ========== List container ========== */
 .pinup_list_wrap {
     width: 100%;
     margin: 0 auto;
-    padding: 1vw;
+    padding: 16px;
     box-sizing: border-box;
 }
 
+/* Flex → Grid, 최소 카드 폭 보장 */
 .pinup_list {
-    display: flex;
-    gap: 1vw;
-    flex-wrap: wrap; /* 아이템들이 한 줄에 꽉 차면 다음 줄로 넘어가도록 설정 */
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+    gap: 16px;
 }
 
+/* ========== Card ========== */
 .pinup_list_wrap .pinup_list_item {
-    flex: 0 0 auto;
-    box-sizing: border-box; /* padding과 margin이 크기 계산에 영향을 미치지 않도록 */
-    width: 13vw;
+    box-sizing: border-box;
+    width: auto;
     background: #ffffff;
     border-radius: 10px;
     overflow: hidden;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    box-shadow: 0 4px 8px rgba(0,0,0,0.1);
     transition: transform 0.3s ease;
-    position: relative;
-    padding-bottom: 3vh;
+    display: flex;
+    flex-direction: column;
+    min-height: 100%;
 }
+.pinup_list_wrap .pinup_list_item:hover { transform: translateY(-5px); }
 
-.pinup_list_wrap .pinup_list_item:hover {
-    transform: translateY(-5px);
-}
-
+/* 이미지 영역: 비율 고정 */
 .pinup_list_item .pinup_img_wrap {
     display: block;
     width: 100%;
-    height: 15vh;
+    aspect-ratio: 4 / 3;
+    height: auto;
     overflow: hidden;
 }
-
 .pinup_list_item .pinup_img_wrap img {
-    width: 100%;
-    height: 100%;
+    width: 100%; height: 100%;
     object-fit: cover;
     transition: transform 0.3s ease;
 }
+.pinup_list_item .pinup_img_wrap img:hover { transform: scale(1.05); }
 
-.pinup_list_item .pinup_img_wrap img:hover {
-    transform: scale(1.05);
-}
-
-.pinup_list_item .pinup_content {
-    padding: 0px 1vh;
-}
-
+/* 카드 본문 */
+.pinup_list_item .pinup_content { padding: 0 12px; }
 .pinup_list_item .pinup_title {
-    font-size: 16px;
-    font-weight: bold;
-    color: #333;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
+    font-size: 16px; font-weight: bold; color: #333;
+    white-space: nowrap; overflow: hidden; text-overflow: ellipsis;
 }
+.pinup_list_item .pinup_location { font-size: 14px; color: #777; }
 
-.pinup_list_item .pinup_location {
-    font-size: 14px;
-    color: #777;
-}
-
+/* 카드 내 버튼(공통) */
 .pinup_list_item button {
-    background: #ffffff;
-    color: #333;
-    border: none;
-    padding: 0px 0px 0px 0px;
-    border-radius: 5px;
-    font-size: 14px;
-    cursor: pointer;
+    background: #fff; color: #333; border: none; padding: 0;
+    border-radius: 5px; font-size: 14px; cursor: pointer;
     transition: background 0.3s ease;
-    margin: 1vh 0.5vh;
+    margin: 4px 6px;
 }
+.pinup_list_item button:hover { background: #ffffff; }
 
-.pinup_list_item button:hover {
-    background: #ffffff;
-}
-.list_pinup_top {
-    display: flex;
-    flex-wrap: wrap;
-    width: 100%;
-}
-.wrap_profile {
-    margin-bottom: 5px;
-}
-
-.post_content_wrap{
-    margin: 0 auto;
-    padding: 20px 0;
-}
-
+/* 닉네임 기본 */
 .pinup_nickname {
-    overflow: hidden;
-    font-size: 1vh;
-    line-height: 1.2;
-    color: #444;
-    text-overflow: ellipsis;
-    word-break: break-word;
-    white-space: nowrap;
-    text-align: left;
-    max-width: 220px;
-    margin: 1vh;
-    -webkit-line-clamp: 1;
+    overflow: hidden; font-size: 12px; line-height: 1.2; color: #444;
+    text-overflow: ellipsis; word-break: break-word; white-space: nowrap;
+    text-align: left; max-width: 220px; margin: 6px; -webkit-line-clamp: 1;
 }
 
-.pinup_list_item .post_action_wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: space-between; /* 버튼과 닉네임 사이 간격 유지 */
-    padding: 0vh 1vh;
-    border-top: 1px solid #eee;
-    border-radius: 0 0 10px 10px; /* 카드 모양과 어울리게 */
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-    box-sizing: border-box;
-}
-
-.post_edit_button,
-.post_delete_button {
-    background: #fff;
-    border: 1px solid #ccc;
-    font-size: 12px;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-}
-
-/* 🔹 본문 내용 스타일 */
+/* 본문 요약 */
 .post_description_wrapper {
-    height: 5vh;
+    height: auto;
+    max-height: calc(1.4em * 3);
     overflow: hidden;
     border-top: 1px solid #eee;
 }
-
-/* 🔹 본문 텍스트 스타일 */
 .post_description_text {
-    height: 100%; /* 부모 높이에 맞춤 */
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: -webkit-box;
-    -webkit-line-clamp: 3; /* 최대 3줄 */
-    -webkit-box-orient: vertical;
-    font-size: 1vh;
-    color: #666;
-    line-height: 1.5;
+    height: 100%;
+    overflow: hidden; text-overflow: ellipsis;
+    display: -webkit-box; -webkit-line-clamp: 3; -webkit-box-orient: vertical;
+    font-size: 12px; color: #666; line-height: 1.4;
 }
 
-/* 🔹 날짜 및 댓글 정보 스타일 */
+/* 날짜/댓글/좋아요 줄 */
 .post_sub_info {
-    display: flex;
-    align-items: center;
-    gap: 1vh;
-    font-size: 13px;
-    color: #777;
+    display: flex; align-items: center;
+    flex-wrap: nowrap; white-space: nowrap;
+    gap: 8px;
+    font-size: 13px; color: #777;
     border-top: 1px solid #eee;
-    padding: 0.9vh 0;
-    margin-bottom: 0.6vh;
+    padding: 8px 0;
+    margin-bottom: 6px;
+    overflow: hidden;
 }
+.comment_meta, .like_meta { display: inline-flex; align-items: center; gap: 4px; }
+.post_date { font-size: 13px; color: #555; }
+.post_separator { color: #aaa; }
 
-.post_date {
-    font-size: 13px;
-    color: #555;
+/* ========== 하단 액션바(닉네임/수정/삭제) ========== */
+.post_action_wrapper {
+    position: static;
+    margin-top: auto;
+    display: flex; align-items: center; justify-content: space-between; gap: 8px;
+    padding: 8px 12px;
+    border-top: 1px solid #eee;
+    border-radius: 0 0 10px 10px;
+    min-height: 40px;
 }
-
-.post_separator {
-    color: #aaa;
+/* 액션바 내부: 닉네임은 좌측 고정 + 한 줄 */
+.post_action_wrapper .pinup_nickname {
+    margin-right: auto; min-width: 0;
+    white-space: nowrap; overflow: hidden; text-overflow: ellipsis;
 }
-
+/* 액션바 버튼: 세로 마진으로 높이 늘어나는 것 방지 */
+.post_edit_button, .post_delete_button {
+    background: #fff; border: 1px solid #ccc; font-size: 12px;
+    border-radius: 5px; cursor: pointer; transition: all 0.3s ease;
+    margin: 0 6px;               /* 세로 0 고정 */
+}