… src/main/resources/templates/views/posts/list.html