## 요약

- **목표:** 게시글 목록 조회 성능 최적화 + 리스트 카드에 **좋아요 UI** 도입 + **메타 영역(날짜/댓글/좋아요)** 줄바꿈 없이 안정 표시 + **REST Docs/테스트 정합**
- **결과:** 단일 요청 **×39.8배** 개선(9.147s → 0.230s), 부하(p95) **×7~8배** 개선, 리스트/상세 간 좋아요 상태 **실시간 동기화**, 문서·테스트를 **JPQL/DTO 구조에 맞게 갱신**

---

## 작업 내용

### 1) 백엔드 성능 최적화 (perf)

- **JPQL 리팩토링:** N+1 제거 → **DTO 프로젝션 + `EXISTS` 서브쿼리**로 단일 조회
- **`GROUP BY` 단순화:** 불필요 컬럼 제거로 플래너 효율 향상
- **Service 로직 정리:** 다중 쿼리 → **Repository 단일 호출**
- **DTO 정합:** `PostResponse` 생성자/필드 구성 JPQL 결과에 맞게 조정
    
    *응답 스키마(필드명)는 기존과 동일*
    

**변경 파일(예시)**

`PostResponse.java`, `PostRepository.java`, `PostService.java`, `MemberResponse.java`

### 2) 프론트 좋아요 UI 도입 (feat)

- **list.html:** 본문 요약 제거, **하트/카운트** 추가, 비로그인 시 **읽기전용** 버튼 처리
- **post.js:** `initializeLikeButtons()`에 **LikeSync.record** 연동, 이벤트 위임/전파 차단, **리스트↔모달** 동기화
- **store-detail.js:** 동적 DOM에 **initializeLikeButtons 재바인딩**, `LikeSync.apply(postContainer)` 적용

**변경 파일**

`templates/views/posts/list.html`, `static/js/post.js`, `static/js/store-detail.js`

### 3) 리스트 레이아웃 안정화 (refactor)

- **그리드 기반 카드 배치**(minmax 280px), 이미지 **aspect-ratio**로 찌그러짐 방지
- 하단 **메타 한 줄 유지**(nowrap, 고정 gap), 댓글/좋아요 숫자 **폭 고정**
- 액션바 **static + min-height: 40px**, 버튼 상하 마진 0, 닉네임 **ellipsis**
- `vw/vh` 의존 축소 → **해상도별 일관된 표시**

**변경 파일**

`static/css/post_list.css`

### 4) 테스트/문서 보강 (test/docs)

- **REST Docs**
    - 목록 응답 문서화: `responseFields` → **`relaxedResponseFields`** 로 변경(누락 필드 허용)
    - 생성/수정/상세 응답에 **`likeCount`** 필드 문서화 추가
    - **선택 필드**로 처리: `storeId`, `member.*`, `content`, `updatedAt` 등
- **단위·슬라이스 테스트**
    - 리포지토리 스텁을 **`findPostListItems(storeId, isDeleted, memberId)`** 로 교체
    - `PostResponse` 생성자 사용 정정

**변경 파일**

`PostApiControllerDocsTest.java`, `PostApiControllerSliceTest.java`, `PostServiceUnitTest.java`

---

## 참고 사항

### 성능 측정 (3편 결과 요약)

- **단일 요청(`curl /post/list/1`)**
    - 인덱스 전: **9.147s**
    - 인덱스 후: **0.450s** (×20.3)
    - 인덱스+JPQL: **0.230s** (×2.0, 누적 ×39.8)
- **k6 Baseline (1VU, 1분) p95**
    - 304.43ms → 70.48ms → **41.34ms** (누적 ×7.36)
- **k6 Light (10VU, 3분) p95**
    - 320.72ms → 42.33ms → **39.55ms** (누적 ×8.11)

> 메모: 측정은 인덱스(comments.post_id, posts.store_id, posts.created_at) 적용 상태 기준. 운영 DB 반영 여부는 배포 전 확인 바랍니다.
> 
> 
> *(post_like의 `(post_id, member_id)` 복합 인덱스는 현재 보류 — 추후 트래픽 증가 시 도입 검토)*
> 

### 호환성/리스크

- API **응답 스키마 변화 없음**(필드 동일). 내부 JPQL/생성자만 변경.
- 프론트는 비로그인 시 **읽기전용** 처리 + 액션 클릭 시 **401 유도 로직** 정상 동작 전제.
- CSS 변경은 레이아웃 안정화 위주(브레이킹 체인지 없음).
- REST Docs는 **실제 생성된 스니펫 기준**으로 include 필요(요청 바디 없는 GET/DELETE에 `request-fields` include 지양).

### 테스트 포인트

- 비로그인: 리스트/상세에서 좋아요 클릭 시 **로그인 안내/리다이렉트**
- 로그인: 리스트/상세에서 좋아요 토글 → **양쪽 카운트/상태 동기화**
- 목록 API 응답 시간: **샘플링**으로 p95 추적
- 다양한 화면 폭(노트북/데스크톱)에서 **메타 영역 한 줄 유지**

---

## 관련 이슈

- Close #이슈번호